+++ src/Symfony/Component/String/AbstractString.php
@@ -255,6 +255,14 @@ abstract class AbstractString implements \Stringable, \JsonSerializable
     }
 
     /**
+     * @param string|string[] $needle
+     */
+    public function containsAny($needle): bool
+    {
+        return null !== $this->indexOf($needle);
+    }
+
+    /**
      * @param string|string[] $suffix
      */
     public function endsWith($suffix): bool