+++ src/Symfony/Component/String/Tests/AbstractAsciiTestCase.php
@@ -55,6 +55,26 @@ abstract class AbstractAsciiTestCase extends TestCase
         ];
     }
 
+    /**
+     * @dataProvider provideIndexOf
+     */
+    public function testContainsAny(?int $result, string $string, $needle)
+    {
+        $instance = static::createFromString($string);
+
+        $this->assertSame(null !== $instance->indexOf($needle), $instance->containsAny($needle));
+    }
+
+    /**
+     * @dataProvider provideIndexOfIgnoreCase
+     */
+    public function testContainsAnyIgnoreCase(?int $result, string $string, $needle)
+    {
+        $instance = static::createFromString($string);
+
+        $this->assertSame(null !== $instance->ignoreCase()->indexOf($needle), $instance->ignoreCase()->containsAny($needle));
+    }
+
     public function testUnwrap()
     {
         $expected = ['hello', 'world'];
@@ -161,7 +181,7 @@ abstract class AbstractAsciiTestCase extends TestCase
     /**
      * @dataProvider provideIndexOf
      */
-    public function testIndexOf(?int $result, string $string, string $needle, int $offset)
+    public function testIndexOf(?int $result, string $string, $needle, int $offset)
     {
         $instance = static::createFromString($string);
 
@@ -180,6 +200,7 @@ abstract class AbstractAsciiTestCase extends TestCase
             [null, 'abc', 'a', -1],
             [null, '123abc', 'B', -3],
             [null, '123abc', 'b', 6],
+            [0, 'abc', ['a', 'e'], 0],
             [0, 'abc', 'a', 0],
             [1, 'abc', 'b', 1],
             [2, 'abc', 'c', 1],
@@ -191,7 +212,7 @@ abstract class AbstractAsciiTestCase extends TestCase
     /**
      * @dataProvider provideIndexOfIgnoreCase
      */
-    public function testIndexOfIgnoreCase(?int $result, string $string, string $needle, int $offset)
+    public function testIndexOfIgnoreCase(?int $result, string $string, $needle, int $offset)
     {
         $instance = static::createFromString($string);
 
@@ -208,6 +229,7 @@ abstract class AbstractAsciiTestCase extends TestCase
             [null, 'abc', 'a', -1],
             [null, 'abc', 'A', -1],
             [null, '123abc', 'B', 6],
+            [0, 'ABC', ['a', 'e'], 0],
             [0, 'ABC', 'a', 0],
             [0, 'ABC', 'A', 0],
             [1, 'ABC', 'b', 0],